requirements.txt
pandas
numpy
scikit-learn
nltk
spacy
pdfplumber
python-docx
streamlit
openpyxl
